Fix multi-region pruning.
The previous code had a single static function for deleting snapshots,
which depends on a global `var e2` from the outer scope. The problem is
that this `e2` variable keeps getting overridden as we map over the
possible regions. As a result, by the time we have finished querying for
a list of snapshots, the `ec2` object likely now points to a version
from a different EC2 region, and the `deleteSnapshot` call fails with:

> InvalidSnapshot.NotFound: The snapshot 'snap-04b012caaa430aab0' does
> not exist.

The solution is to curry the `delOneSnap` function, wrapping it in
another function which takes the `ec2` object as an argument. The
resulting inner function is the thing which is actually provided to
`mapSeries`, and this inner function no longer depends on the global
`ec2`, instead reading from the `ec2` parameter from the outer function.